Move some functions from the 'float' implementation into a separate file
This is in preparation for a s2f implememtation.
No functional change.
